uint32 in db to uint64 because sqlite INTEGER are int64 atk and def can be negative so int64 route to get a single card

+++ models/entity.card.go
@@ -1,17 +1,17 @@
 package model
 
 type EntityCard struct {
-	ID        uint32 `gorm:"type:INTEGER PRIMARYKEY"`
-	Ot        uint32 `gorm:"type:INTEGER"`
-	Alias     uint32 `gorm:"type:INTEGER"`
-	SetCode   uint32 `gorm:"type:INTEGER"`
-	Type      uint32 `gorm:"type:INTEGER"`
-	Atk       uint32 `gorm:"type:INTEGER"`
-	Def       uint32 `gorm:"type:INTEGER"`
-	Level     uint32 `gorm:"type:INTEGER"`
-	Race      uint32 `gorm:"type:INTEGER"`
-	Attribute uint32 `gorm:"type:INTEGER"`
-	Category  uint32 `gorm:"type:INTEGER"`
+	ID        uint64 `gorm:"type:INTEGER PRIMARYKEY"`
+	Ot        uint64 `gorm:"type:INTEGER"`
+	Alias     uint64 `gorm:"type:INTEGER"`
+	SetCode   uint64 `gorm:"type:INTEGER"`
+	Type      uint64 `gorm:"type:INTEGER"`
+	Atk       int64  `gorm:"type:INTEGER"`
+	Def       int64  `gorm:"type:INTEGER"`
+	Level     uint64 `gorm:"type:INTEGER"`
+	Race      uint64 `gorm:"type:INTEGER"`
+	Attribute uint64 `gorm:"type:INTEGER"`
+	Category  uint64 `gorm:"type:INTEGER"`
 	Name      string `gorm:"type:TEXT"`
 	Desc      string `gorm:"type:TEXT"`
 }
